updated chart render to support OADP stable-1.5 channel for OCP 4.19 (#2147)
